simulator/requirements.txt
fastapi
uvicorn[standard]
structlog
docker
requests
pyyaml
